requirements.txt
paddlepaddle
paddleocr
easyocr
surya-ocr
pytesseract
Pillow
opencv-python
python-Levenshtein
